Added BIGINT for order_id

+++ data-source-layer/db/DatabaseScript.sql
@@ -527,7 +527,7 @@ CREATE TABLE CART (
 
 /* ------------------------------------------ ORDER TABLE QUERIES --------------------------------------------- */
 CREATE TABLE ORDERS (
-    order_id INTEGER NOT NULL UNIQUE,
+    order_id BIGINT NOT NULL UNIQUE,
     user_id SERIAL REFERENCES ACTIVEUSERS(user_id) NOT NULL,
     orderDate DATE NOT NULL default CURRENT_DATE,
     total DECIMAL NOT NULL DEFAULT 0,
@@ -539,7 +539,7 @@ CREATE TABLE ORDERS (
 
 CREATE TABLE ORDERITEM (
     order_item_id VARCHAR(60) NOT NULL UNIQUE,
-    order_id INTEGER REFERENCES ORDERS(order_id) NOT NULL,
+    order_id BIGINT REFERENCES ORDERS(order_id) NOT NULL,
     serialNumber VARCHAR(10) NOT NULL,
     price DECIMAL NOT NULL,
     isReturned BOOLEAN DEFAULT FALSE,
@@ -550,7 +550,7 @@ CREATE TABLE ORDERITEM (
 
 CREATE TABLE RETURNS (
     return_id SERIAL PRIMARY KEY NOT NULL,
-    order_id INTEGER REFERENCES ORDERS(order_id) NOT NULL,
+    order_id BIGINT REFERENCES ORDERS(order_id) NOT NULL,
     user_id SERIAL REFERENCES ACTIVEUSERS(user_id) NOT NULL,
     order_item_id VARCHAR(60) REFERENCES ORDERITEM(order_item_id) NOT NULL
 );
